完善matomo数据上报

+++ android/app/src/main/java/com/datizhuanqian/matomo/MatomoModule.java
@@ -9,8 +9,8 @@ import org.matomo.sdk.Matomo;
 import org.matomo.sdk.Tracker;
 import org.matomo.sdk.TrackerBuilder;
 import org.matomo.sdk.extra.TrackHelper;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -46,7 +46,7 @@ public class MatomoModule extends ReactContextBaseJavaModule implements Lifecycl
     }
 
     @ReactMethod
-    public void setDispatchInterval(long interval) {
+    public void setDispatchInterval(int interval) {
         mMatomoTracker.setDispatchInterval(interval);
     }
